New layout
For some reason, the CEO thinks we need a new design, so we are
redesigning a perfectly good site, and cluttering up our code, since I
couldn't manage to properly adapt it with the templates style.

We still need to properly organize our new color scheme.

And the `process-slider` is still a Work in Progress.